django-portal/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
Django==3.2.7
djangorestframework==3.12.4
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
pyproj==3.2.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
response==0.5.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
